requirements.txt
torch==2.7.1
torchmetrics==0.10.3
pytorch-lightning==2.5.2
transformers==4.53.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
fasttext==0.9.3
entmax==1.3
scikit-learn==1.7.0
scipy==1.16.0
pandas==2.3.1
numpy==1.26.4
tqdm==4.67.1
datasets==3.6.0
huggingface-hub==0.33.4
safetensors==0.5.3
protobuf==4.25.8
pyarrow==20.0.0
lxml==6.0.0
unbabel-comet==2.2.6
sacrebleu==2.5.1
requests
pyfranc
polyglot
lingua-language-detector
PyICU

wget==3.2
tabulate==0.9.0
pyyaml==6.0.2
